local-voice-control/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
azure-cognitiveservices-speech>=1.35.0
openai>=1.10.0
